[Silver III] Title: 주유소, Time: 24 ms, Memory: 3588 KB, Score: 100 point -BaekjoonHub

+++ 백준/Silver/13305. 주유소/주유소.cc
@@ -21,8 +21,8 @@ int main(){
         cin >> prices[i];
     }
     
-    int result = 0;
-    int min_prices = prices[0];
+    long long result = 0;
+    long long min_prices = prices[0];
     
     for(int i=0;i<N-1;i++) {
         if(i==0){
